Move FixPathCase to Common/File/Path
